refactor: Report 엔티티 컨트롤러/서비스 분리 및 api url 수정
